fix backend test
